Support reset_peer toxic
Besides adding the new toxic, this commit removes a number of
deprecated libraries and tools from the project.

The `requests` library was replaced by `axios` which is actively maintained.

The development tools used in this project were also outdated.
`tslint` was replaced by `eslint` and `ava` was replaced by `jest`.

The `tslint` migration as well as the `ava` migration were conducted
mostly automatic with only a few manual modifications.