feat(api): make `Proxy` and `Toxic` classes immutable and remove cyclic dependencies (#10)
Removing the cyclic dependencies between the objects has many benefits,
for example JSON (de-)serialization of Axios errors now just works.

BREAKING CHANGE: `Proxy` and `Toxic` objects are now updated
by posting an update to the server with the attributes to be
updated. This is a breaking change because the `update` methods
return a new instance of a `Proxy` or `Toxic` object instead of
returning the mutated object.